requirements.txt
ffmpeg-python==0.2.0
pyttsx3>=2.90
requests>=2.26.0
openai>=0.27.0
moviepy>=1.0.3
numpy>=1.21.0
Pillow>=8.3.1
python-dotenv>=0.19.0
elevenlabs>=0.2.0
pydub>=0.25.1
Jinja2==3.1.2
PyYAML==6.0
pydantic==1.10.2
Flask==2.2.2
APScheduler==3.9.1.post1
aiohttp==3.8.4
loguru==0.6.0
beautifulsoup4>=4.9.3
lxml>=4.9.0

# Optional dependencies for advanced features
# google-cloud-texttospeech>=2.11.0
# boto3>=1.26.0
